refactor: move storeOriginalSearchUrl utility to a dedicated file
